requirements.txt
lxml==4.0.0
tldextract==2.1.0
Unidecode==0.04.21
pymongo==3.5.1
selenium==3.6.0
python_dateutil==2.6.1
